requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=0.6,!=0.7,<1.0
#
# The greenlet package must be compiled with gcc and needs
# the Python.h headers. Make sure you install the python-dev
# package to get the right headers...
greenlet>=0.3.2
eventlet>=0.15.1
WebOb
oslo.config
oslo.utils
oslo.log
oslo.i18n
oslo.serialization
oslo.concurrency
